+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 5.4.8 (unreleased)
 ------------------
 
+
 - Handle None values in link integrity blocks adapter.
+- Patch in @querystring-search that avoid to search through all the site if there is an absolutePath criteria with non existing UID and b_size==1.
+  See #99 for more details.
   [cekk]